lc-practices/4-langserve/requirements.txt
python-dotenv

langchain
langchain-community
langchain-openai

fastapi
uvicorn
langserve 
sse_starlette
pydantic